stocks_recommendation/requirements.txt
pyyaml
pandas
selenium
pytest
pymysql
sqlalchemy
mysqlclient
google-api-python-client
google-auth-oauthlib